+++ src/main/java/uk/gov/hmcts/reform/divorce/fees/models/Fee.java
@@ -18,4 +18,6 @@ public class Fee {
     double amount;
 
     int  version;
+
+    String description;
 }
