+++ src/main/java/uk/gov/hmcts/reform/divorce/pay/exceptions/PaymentFailedException.java
@@ -1,0 +1,9 @@
+package uk.gov.hmcts.reform.divorce.pay.exceptions;
+
+public class PaymentFailedException extends RuntimeException{
+
+    public PaymentFailedException(String message) {
+        super(message);
+    }
+
+}
