+++ src/main/java/uk/gov/hmcts/reform/divorce/transformservice/domain/model/ccd/CoreCaseData.java
@@ -427,5 +427,9 @@ public class CoreCaseData {
 
     @JsonProperty("D8SelectedDivorceCentreSiteId")
     private String d8SelectedDivorceCentreSiteId;
+
+    @JsonProperty("PetitionerSolicitorFirm")
+    private String petitionerSolicitorFirm;
+
 }
