+++ src/test/java/uk/gov/hmcts/reform/divorce/transformservice/controller/CcdCallbackControllerTest.java
@@ -4,7 +4,6 @@ import org.apache.commons.io.FileUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -20,7 +19,7 @@ import uk.gov.hmcts.reform.divorce.CaseProgressionApplication;
 import uk.gov.hmcts.reform.divorce.fees.models.Fee;
 import uk.gov.hmcts.reform.divorce.fees.services.FeesAndPaymentService;
 import uk.gov.hmcts.reform.divorce.notifications.service.EmailService;
-import uk.gov.hmcts.reform.divorce.pay.service.PaymentService;
+import uk.gov.hmcts.reform.divorce.pay.services.PaymentService;
 import uk.gov.hmcts.reform.divorce.testutils.ObjectMapperTestUtil;
 import uk.gov.hmcts.reform.divorce.transformservice.client.pdf.PdfGeneratorException;
 import uk.gov.hmcts.reform.divorce.transformservice.domain.ccd.CreateEvent;
